requirements.txt
Flask
Flask-Cors
requests~=2.32.5
python-dotenv~=1.1.1
google-generativeai